update item prices

+++ src/main/java/com/makechi/invizio/controller/metrics/MetricsController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -24,7 +25,7 @@ public class MetricsController {
     }
 
     @GetMapping("/inventory-valuation")
-    public double getInventoryValuation() {
+    public BigDecimal getInventoryValuation() {
         return metricService.calculateInventoryValuation();
     }
