+++ src/main/java/com/makechi/invizio/service/metrics/MetricService.java
@@ -4,12 +4,13 @@ import com.makechi.invizio.dto.metrics.MetricsDto;
 import com.makechi.invizio.dto.metrics.StockLevelDto;
 import org.bson.Document;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface MetricService {
     MetricsDto getAllMetrics();
 
-    double calculateInventoryValuation();
+    BigDecimal calculateInventoryValuation();
 
     List<Document> getItemsGroupedByCategory();
